moved translator logic into its own file
